requirements.txt
Django==3.2.5
beautifulsoup4==4.9.3
requests==2.26.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
